+++ tests/test_admin_api.py
@@ -1,8 +1,16 @@
 from datetime import datetime
+from pathlib import Path
+from zipfile import ZipFile
 
+import pytest
 from devtools import debug
 
 from fractal_server.app.models import JobStatusType
+from fractal_server.app.runner import _backends
+from fractal_server.app.runner._common import SHUTDOWN_FILENAME
+
+backends_available = list(_backends.keys())
+
 
 PREFIX = "/admin"
 
@@ -452,3 +460,120 @@ async def test_patch_job(
         res = await client.get(f"/api/v1/project/{project.id}/job/{job.id}/")
         assert res.status_code == 200
         assert res.json()["status"] == NEW_STATUS
+
+
+@pytest.mark.parametrize("backend", backends_available)
+async def test_stop_job(
+    backend,
+    MockCurrentUser,
+    project_factory,
+    dataset_factory,
+    workflow_factory,
+    job_factory,
+    task_factory,
+    registered_superuser_client,
+    db,
+    tmp_path,
+    override_settings_factory,
+):
+    override_settings_factory(FRACTAL_RUNNER_BACKEND=backend)
+
+    async with MockCurrentUser(user_kwargs={"id": 1111}) as user:
+        project = await project_factory(user)
+        workflow = await workflow_factory(project_id=project.id)
+        task = await task_factory(name="task", source="source")
+        await workflow.insert_task(task_id=task.id, db=db)
+        dataset = await dataset_factory(project_id=project.id)
+        job = await job_factory(
+            working_dir=tmp_path.as_posix(),
+            project_id=project.id,
+            input_dataset_id=dataset.id,
+            output_dataset_id=dataset.id,
+            workflow_id=workflow.id,
+            status=JobStatusType.RUNNING,
+        )
+
+    async with MockCurrentUser(user_kwargs={"id": 2222, "is_superuser": True}):
+
+        res = await registered_superuser_client.get(
+            f"{PREFIX}/job/{job.id}/stop/",
+        )
+
+        if backend == "slurm":
+            assert res.status_code == 204
+            shutdown_file = tmp_path / SHUTDOWN_FILENAME
+            debug(shutdown_file)
+            assert shutdown_file.exists()
+            res = await registered_superuser_client.get(
+                f"{PREFIX}/job/{job.id + 42}/stop/",
+            )
+            assert res.status_code == 404
+        else:
+            assert res.status_code == 422
+
+
+async def test_download_job_logs(
+    client,
+    MockCurrentUser,
+    project_factory,
+    dataset_factory,
+    workflow_factory,
+    job_factory,
+    task_factory,
+    db,
+    tmp_path,
+):
+    async with MockCurrentUser(user_kwargs={"id": 1111}) as user:
+        prj = await project_factory(user)
+        input_dataset = await dataset_factory(project_id=prj.id, name="input")
+        output_dataset = await dataset_factory(
+            project_id=prj.id, name="output"
+        )
+        workflow = await workflow_factory(project_id=prj.id)
+        task = await task_factory()
+        await workflow.insert_task(task_id=task.id, db=db)
+        working_dir = (tmp_path / "workflow_dir_for_zipping").as_posix()
+        job = await job_factory(
+            project_id=prj.id,
+            workflow_id=workflow.id,
+            working_dir=working_dir,
+            input_dataset_id=input_dataset.id,
+            output_dataset_id=output_dataset.id,
+        )
+
+    async with MockCurrentUser(user_kwargs={"id": 2222, "is_superuser": True}):
+
+        # Write a log file in working_dir
+        LOG_CONTENT = "This is a log\n"
+        LOG_FILE = "log.txt"
+        Path(working_dir).mkdir()
+        with (Path(working_dir) / LOG_FILE).open("w") as f:
+            f.write(LOG_CONTENT)
+
+        # Test 404
+        res = await client.get(f"{PREFIX}/job/{job.id + 42}/download/")
+        assert res.status_code == 404
+
+        # Test that the endpoint returns a list with the new job
+        res = await client.get(f"{PREFIX}/job/{job.id}/download/")
+        assert res.status_code == 200
+        assert (
+            res.headers.get("content-type") == "application/x-zip-compressed"
+        )
+
+        # Write response into a zipped file
+        zipped_archive_path = tmp_path / "logs.zip"
+        debug(zipped_archive_path)
+        with zipped_archive_path.open("wb") as f:
+            f.write(res.content)
+
+        # Unzip the log archive
+        unzipped_archived_path = tmp_path / "unzipped_logs"
+        debug(unzipped_archived_path)
+        with ZipFile(zipped_archive_path, mode="r") as zipfile:
+            zipfile.extractall(path=unzipped_archived_path)
+
+        # Verify content of the unzipped log archive
+        with (unzipped_archived_path / LOG_FILE).open("r") as f:
+            actual_logs = f.read()
+        assert LOG_CONTENT in actual_logs