+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.4.1 (unreleased)
 
 * API:
+    * Add `GET /admin/job/{job_id}/stop/` and `GET /admin/job/{job_id}/download/` endpoints (\#1059).
     * Use `DatasetRead` and `WorkflowRead` models for "dump" attributes of `ApplyWorkflowRead` (\#1049).
 * Database:
     * Make `ApplyWorkflow.workflow_dump` column non-nullable (\#1049).